+++ src/main/java/com/range/rpms/dto/user/UserRegisterRequest.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class UserRegisterRequest {
-    private String id;
+
     private String username;
     private String password;
     private String email;