+++ src/main/java/com/range/rpms/exception/GlobalExceptionHandler.java
@@ -1,6 +1,6 @@
 package com.range.rpms.exception;
 
-import com.range.rpms.dto.ApiResponse;
+import com.range.rpms.dto.GenericResponse;
 import com.range.rpms.util.ErrorResponseUtil;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 @ControllerAdvice
 public class GlobalExceptionHandler {
     @ExceptionHandler(AbstractExceptionHandler.class)
-    public ResponseEntity<ApiResponse<Void>> handleException(AbstractExceptionHandler e) {
+    public ResponseEntity<GenericResponse<Void>> handleException(AbstractExceptionHandler e) {
 
         return ErrorResponseUtil
                 .build(e.getHttpStatus()