+++ src/main/java/com/range/rpms/exception/pkg/CantDownGradePackageException.java
@@ -1,7 +1,6 @@
 package com.range.rpms.exception.pkg;
 
 import com.range.rpms.exception.AbstractExceptionHandler;
-import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpStatus;
 
 public class CantDownGradePackageException extends AbstractExceptionHandler {