+++ src/main/java/com/range/rpms/security/SecurityConfig.java
@@ -1,20 +1,13 @@
 package com.range.rpms.security;
 
 import com.range.rpms.security.filter.JwtFilter;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.http.HttpMethod;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.provisioning.InMemoryUserDetailsManager;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 
@@ -22,7 +15,7 @@ import org.springframework.security.web.authentication.UsernamePasswordAuthentic
 @EnableWebSecurity
 public class SecurityConfig {
 
-    private JwtFilter jwtFilter;
+    private final JwtFilter jwtFilter;
 
     public SecurityConfig(JwtFilter jwtFilter) {
         this.jwtFilter = jwtFilter;
@@ -32,15 +25,11 @@ public class SecurityConfig {
 
     @Bean
     public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-http
+        http
                 .csrf(AbstractHttpConfigurer::disable)
                 .addFilterBefore(jwtFilter, UsernamePasswordAuthenticationFilter.class)
                 .formLogin(form -> form
-                        .loginPage("/login") //With get return login.html
-                        .loginProcessingUrl("/perform_login")
-                        .defaultSuccessUrl("/", true)
-                        .failureUrl("/login?error=true")
-                        .permitAll()
+                        .disable()
                 )
                 .logout(logout -> logout
                         .logoutUrl("/logout")
@@ -48,11 +37,28 @@ http
                         .permitAll()
                 )
                 .authorizeHttpRequests(auth -> auth
-                        .requestMatchers("/login", "/css/**", "/js/**", "/images/**").permitAll()
 
-                        .requestMatchers(HttpMethod.DELETE,"/packages/**")
-                        .hasAuthority("ADMIN")
+                        .requestMatchers("/login",
+                                "/swagger",
+                                "v3/api-docs",
+                                "/css/**",
+                                "/js/**",
+                                "/images/**",
+                                "/register",
+                                "auth/register",
+                                "auth/login"
+                        ).permitAll()
 
+                        .requestMatchers("/info/all",
+                                "packages/delete/all",
+                                "packages/delete/"
+                                //for now l set it admin but l will change it in new version
+
+                        ).hasRole("ADMIN")
+
+
+
+.requestMatchers("/").hasRole("USER")
 
 
                         .anyRequest().permitAll()